experemental connection string in cpurepository

+++ TaskAPI-2-1/MetricsAgent/DAL/DotNetMetricsRepository.cs
@@ -44,10 +44,10 @@ namespace MetricsAgent.DAL
                     // добавляем объект в список возврата
                     returnList.Add(new DotNetMetric
                     {
-                        Id = reader.GetInt32(0),
-                        Value = reader.GetInt32(1),
+                        Id = (int)reader.GetInt64(0),
+                        Value = (int)reader.GetInt64(1),
                         // налету преобразуем прочитанные секунды в метку времени
-                        Time = DateTimeOffset.FromUnixTimeSeconds(reader.GetInt32(2))
+                        Time = DateTimeOffset.FromUnixTimeSeconds((int)reader.GetInt64(2))
                     });
                 }
             }
